Replace old resolver with database resolver
The database resolver will resolve all references at once by using a database of
minimal partial paths.
