+++ crates/solidity/testing/sanctuary/src/tests.rs
@@ -187,7 +187,8 @@ fn run_bindings_check(
     output: &ParseOutput,
 ) -> Result<Vec<UnresolvedReference>> {
     let mut unresolved = Vec::new();
-    let binding_graph = create_bindings(version, source_id, output)?;
+    let mut binding_graph = create_bindings(version, source_id, output)?;
+    binding_graph.resolve();
 
     for reference in binding_graph.all_references() {
         if reference.get_file().is_system() {
